Move checksum_algorithm method to Checksum impl, remove panic for edge case
