+++ mountpoint-s3-client/src/mock_client.rs
@@ -256,7 +256,7 @@ impl MockClient {
                     etag: object.etag.as_str().to_string(),
                     storage_class: object.storage_class.clone(),
                     restore_status: object.restore_status,
-                    checksum_algorithm: object.checksum_algorithm(),
+                    checksum_algorithm: object.checksum.checksum_algorithm(),
                 });
             }
         }
@@ -318,7 +318,7 @@ impl MockClient {
                     etag: object.etag.as_str().to_string(),
                     storage_class: object.storage_class.clone(),
                     restore_status: object.restore_status,
-                    checksum_algorithm: object.checksum_algorithm(),
+                    checksum_algorithm: object.checksum.checksum_algorithm(),
                 });
             }
             next_continuation_token += 1;
@@ -473,24 +473,6 @@ impl MockObject {
 
     pub fn etag(&self) -> ETag {
         self.etag.clone()
-    }
-
-    pub fn checksum_algorithm(&self) -> Option<ChecksumAlgorithm> {
-        let Checksum {
-            checksum_crc32,
-            checksum_crc32c,
-            checksum_sha1,
-            checksum_sha256,
-        } = &self.checksum;
-
-        match (checksum_crc32, checksum_crc32c, checksum_sha1, checksum_sha256) {
-            (Some(_), None, None, None) => Some(ChecksumAlgorithm::Crc32),
-            (None, Some(_), None, None) => Some(ChecksumAlgorithm::Crc32c),
-            (None, None, Some(_), None) => Some(ChecksumAlgorithm::Sha1),
-            (None, None, None, Some(_)) => Some(ChecksumAlgorithm::Sha256),
-            (None, None, None, None) => None,
-            _ => unreachable!("should not have more than one checksum"),
-        }
     }
 }
